Add isBoom flag for tracking of Errors
`instanceof Error` will fail if errors cross over the VM boundary, are wrapped and then are passed back. As there is no clean way to marshal these, adding a `isBoom` flag to allow both sides of the world to operate cleanly.

It's recommended that all code that needs to detect if a particular object is an error do `err.isBoom || err instanceof Error` to properly check for both possible cases.