Allow for pool queue limits
Adds `queueTimeout` and `maxQueue` pool options which can be used to help prevent runaway queues and allow active failover when a pool has entered queuing and partial starvation.

The general intent here is to allow for requests to preemptively serve client side rendering responses when the server side path is under stress.